Split oversized runtime snapshot regression files
