+++ .pm/tasks/task_494b9cd1a63544bca3c1e69f17af1918.execution.md
@@ -276,3 +276,11 @@ Example:
 - Review Findings Disposition: no_findings
 - Finding Disposition Evidence: no code changes were required after the fresh runtime/QA review; this packet binds the reviewed code head to the review results, and any later source-head change before PR creation is limited to `.pm/tasks/task_494b9cd1a63544bca3c1e69f17af1918.{yaml,execution.md}` evidence files only.
 - Residual Risk: `runtime_snapshot_advances_during_slow_gap_sync_before_poll_finishes` is a timing-sensitive regression that may transiently time out under cargo lock contention; the new observer on `192.168.1.3` is alive and correctly catching up but is not yet `ready`; and current verification is focused plus remote deploy proof rather than one fresh full cross-platform regression pass.
+
+## 2026-06-10 00:18:23 CST / tpm
+- 完成内容: 继续执行 PR #395 的 post-PR CI 修复循环，已确认当前唯一 actionable failure 是 GitHub Actions run `27217590455` 的 `required-gate`，原因不是行为回退而是 Rust 文件行数门禁；同时在本地重跑 focused regressions 时，发现 `runtime_snapshot_advances_during_slow_gap_sync_before_poll_finishes` 的前置 sequencer 启动步骤会在共享 cargo/build 锁竞争下偶发超时。
+- Action: 1) 保持生产修复逻辑不变，仅做结构拆分以通过 `check-rust-file-size.sh`；2) 把 `tick` / `tick_with_progress` 移入 `node_engine_core_tick_impl.rs`，把进度相关回归移入 `tests_storage_replication_progress.rs`；3) 将 `runtime_snapshot_advances_during_slow_gap_sync_before_poll_finishes` 的前置数据源改成 deterministic replication history + registered fetch handlers，继续验证同一 observer gap-sync 中途 snapshot 发布行为，但去掉对独立 `runtime_a` 在限定秒数内先跑高的脆弱依赖。
+- Validation Command: `gh pr view 395 --repo eng-cc/oasis7 --json mergeStateStatus,mergeable,reviewDecision,statusCheckRollup`; `gh run view 27217590455 --repo eng-cc/oasis7 --job 80363359666 --log`; `rtk ./scripts/check-rust-file-size.sh`; `rtk env -u RUSTC_WRAPPER cargo test -p oasis7_node runtime_snapshot_advances_during_slow_gap_sync_before_poll_finishes -- --nocapture`; `rtk env -u RUSTC_WRAPPER cargo test -p oasis7_node gap_sync_limits_single_poll_work_to_publish_intermediate_progress -- --nocapture`; `rtk env -u RUSTC_WRAPPER cargo test -p oasis7_node pos_engine_snapshot_reflects_synced_replication_progress_after_record -- --nocapture`; `rtk git diff --check`
+- Expected Result: 结构拆分后 file-size gate 通过；三条 focused regressions 在当前机器负载下稳定通过；PR 可推送新 head 触发 fresh checks，而无需再解释为行为层回退。
+- Actual Result: in progress.
+- Blocker / Next Action: 重跑本地验证；若通过则提交并 push 到 `codex/publish-runtime-snapshot-observer-gap-sync`，随后继续盯 PR #395 checks / comments / mergeability。
